api/requirements.txt
aniso8601==9.0.1
blinker==1.8.0
click==8.1.7
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
uWSGI==2.0.25.1
Werkzeug==3.0.2
